Add files via upload
Eye Blink Detector basically the idea is very simple here we're using cvzone's faceMeshModule to detect the landmarks of the eye of the user. After the landmarks are detected successfully we chose 4 landmarks on both the eyes(one in the EyeBlinkCount and EyeBlinkCountV2) to detect whether the person is blinking or not. This project has various applications in Medical field like detection of Parkinson's Disease or Patient Monitoring etc. it can also be implemented with a drowsiness detection project to detect whether the driver is feeling sleepy or not which can effectively prevent accidents